Updated project to structure from package-by-domain to package-by-feature
