+++ verein/nbproject/project.properties
@@ -106,4 +106,5 @@ run.test.classpath=\
     ${build.test.classes.dir}
 source.encoding=UTF-8
 src.dir=src
+src.resources.dir=resources
 test.src.dir=test
