+++ verein/nbproject/project.xml
@@ -5,6 +5,7 @@
         <data xmlns="http://www.netbeans.org/ns/j2se-project/3">
             <name>verein</name>
             <source-roots>
+                <root id="src.resources.dir" name="Resources"/>
                 <root id="src.dir"/>
             </source-roots>
             <test-roots>
