+++ verein/src/de/lebk/verein/club/ClubManager.java
@@ -1,0 +1,16 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package de.lebk.verein.club;
+
+import javax.swing.JPanel;
+
+/**
+ *
+ * @author sopaetzel
+ */
+public class ClubManager extends JPanel {
+
+}
