--- verein/src/de/lebk/verein/view/component_secondlevel/ClubManager.java
@@ -1,16 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package de.lebk.verein.view.component_secondlevel;
-
-import javax.swing.JPanel;
-
-/**
- *
- * @author sopaetzel
- */
-public class ClubManager extends JPanel {
-
-}
