--- verein/src/de/lebk/verein/view/component_thirdlevel/ImageLabel.java
@@ -1,53 +1,0 @@
-/*
- * The MIT License
- *
- * Copyright 2016 sopaetzel.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package de.lebk.verein.view.component_thirdlevel;
-
-import java.awt.Graphics;
-import java.awt.image.BufferedImage;
-import java.io.File;
-import java.io.IOException;
-import java.net.MalformedURLException;
-import javax.imageio.ImageIO;
-import javax.swing.ImageIcon;
-import javax.swing.JLabel;
-import javax.swing.JPanel;
-
-/**
- *
- * @author sopaetzel
- */
-//TODO: Change to LABEL
-public class ImageLabel extends JLabel {
-
-    private BufferedImage image;
-    private BufferedImage scaledImage;
-
-    public ImageLabel(String imagePath) throws IOException {
-        image = ImageIO.read(new File(imagePath));
-        int ratio = image.getWidth() / image.getHeight();
-        scaledImage = new BufferedImage(100, 86, BufferedImage.TYPE_INT_ARGB);
-        scaledImage.getGraphics().drawImage(image, 0, 0, 100, 86, null);
-        this.setIcon(new ImageIcon(scaledImage));
-    }
-}
